LayoutTests/fast/images/text-recognition/ios/selection-rects-in-image-overlay-expected.txt

PASS initialSelectionRects.length is 1
PASS selectionRectsAfterSelectingAll.length is 4
PASS successfullyParsed is true

TEST COMPLETE
